requirements.txt
langchain
pypdf
faiss-cpu
sentence-transformers
streamlit
PyYAML
fastapi
uvicorn
openai
langchain-openai 